iam_grant_function/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-resource-manager==1.*
